Move the database and repository packages to a subpackage data.
