+++ alembic.ini
@@ -60,7 +60,7 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-# We use the db connection defined in slackhealthbot.database.connection
+# We use the db connection defined in slackhealthbot.data.database.connection
 #sqlalchemy.url = driver://user:pass@localhost/dbname
