+++ slackhealthbot/domain/modelmappers/repositorytodomain/sleep.py
@@ -1,5 +1,5 @@
+from slackhealthbot.data.repositories import fitbitrepository
 from slackhealthbot.domain.models.sleep import SleepData
-from slackhealthbot.repositories import fitbitrepository
 
 
 def repository_sleep_to_core_sleep(
