+++ slackhealthbot/domain/usecases/fitbit/usecase_get_last_sleep.py
@@ -2,12 +2,12 @@ import datetime
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from slackhealthbot.data.repositories import fitbitrepository
 from slackhealthbot.domain.modelmappers.remoteservicetodomain.sleep import (
     remote_service_sleep_to_core_sleep,
 )
 from slackhealthbot.domain.models.sleep import SleepData
 from slackhealthbot.remoteservices.fitbit import sleepapi
-from slackhealthbot.repositories import fitbitrepository
 
 
 async def do(
