+++ slackhealthbot/domain/usecases/fitbit/usecase_post_user_logged_out.py
@@ -1,9 +1,9 @@
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from slackhealthbot.data.repositories import fitbitrepository
 from slackhealthbot.domain.usecases.slack import (
     usecase_post_user_logged_out as slack_usecase_post_user_logged_out,
 )
-from slackhealthbot.repositories import fitbitrepository
 
 
 async def do(
