+++ slackhealthbot/domain/usecases/withings/usecase_process_new_weight.py
@@ -1,9 +1,9 @@
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from slackhealthbot.data.repositories import withingsrepository
 from slackhealthbot.domain.models.weight import WeightData
 from slackhealthbot.domain.usecases.slack import usecase_post_weight
 from slackhealthbot.domain.usecases.withings import usecase_get_last_weight
-from slackhealthbot.repositories import withingsrepository
 
 
 async def do(
