+++ slackhealthbot/routers/dependencies.py
@@ -1,6 +1,6 @@
 from fastapi.templating import Jinja2Templates
 
-from slackhealthbot.database.connection import SessionLocal, ctx_db
+from slackhealthbot.data.database.connection import SessionLocal, ctx_db
 
 
 async def get_db():
