+++ slackhealthbot/scheduler.py
@@ -6,14 +6,14 @@ import logging
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from slackhealthbot.core.exceptions import UserLoggedOutException
-from slackhealthbot.database.connection import SessionLocal
+from slackhealthbot.data.database.connection import SessionLocal
+from slackhealthbot.data.repositories import fitbitrepository
+from slackhealthbot.data.repositories.fitbitrepository import UserIdentity
 from slackhealthbot.domain.usecases.fitbit import (
     usecase_process_new_activity,
     usecase_process_new_sleep,
 )
 from slackhealthbot.domain.usecases.slack import usecase_post_user_logged_out
-from slackhealthbot.repositories import fitbitrepository
-from slackhealthbot.repositories.fitbitrepository import UserIdentity
 from slackhealthbot.settings import settings
